backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.5.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
greenlet>=3.0.0
